Añade shortcut component reusable a nuevo directorio
